minecraft/conn.go: Apply review suggestion.

+++ minecraft/conn.go
@@ -1159,17 +1159,17 @@ func (conn *Conn) handleResourcePackDataInfo(pk *packet.ResourcePackDataInfo) er
 
 	// The client calculates the chunk count by itself: You could in theory send a chunk count of 0 even
 	// though there's data, and the client will still download normally.
-	chunkCount := uint32(pk.Size / uint64(pk.DataChunkSize))
+	chunkCount := int32(pk.Size / uint64(pk.DataChunkSize))
 	if pk.Size%uint64(pk.DataChunkSize) != 0 {
 		chunkCount++
 	}
 
 	idCopy := pk.UUID
 	go func() {
-		for i := uint32(0); i < chunkCount; i++ {
+		for i := int32(0); i < chunkCount; i++ {
 			_ = conn.WritePacket(&packet.ResourcePackChunkRequest{
 				UUID:       idCopy,
-				ChunkIndex: int32(i),
+				ChunkIndex: i,
 			})
 			select {
 			case <-conn.ctx.Done():
